refactor task interface
