+++ src/interfaces/task/BaseTask.ts
@@ -1,0 +1,6 @@
+export interface BaseTask {
+	name: string;
+	start: string;
+	end: string;
+	completion: number | string;
+}
