+++ src/interfaces/task/Task.ts
@@ -1,0 +1,7 @@
+import { BaseTask } from "./BaseTask";
+import { SubTask } from "./SubTask";
+
+export interface Task extends BaseTask {
+	[key: string]: SubTask[] | string | number | undefined;
+	subTasks?: SubTask[];
+}
